+++ src/OpenTelemetry.AutoInstrumentation/Properties/AssemblyInfo.cs
@@ -18,6 +18,8 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("OpenTelemetry.AutoInstrumentation.AspNetCoreBootstrapper")]
+[assembly: InternalsVisibleTo("OpenTelemetry.AutoInstrumentation.Bootstrapping.Tests")]
 [assembly: InternalsVisibleTo("OpenTelemetry.AutoInstrumentation.Tests")]
 [assembly: InternalsVisibleTo("IntegrationTests")]
+[assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
 #endif