+++ src/OpenTelemetry.AutoInstrumentation/Tagging/Tags.cs
@@ -21,14 +21,8 @@ namespace OpenTelemetry.AutoInstrumentation.Tagging;
 /// <summary>
 /// Standard span tags used by integrations.
 /// </summary>
-public static class Tags
+internal static class Tags
 {
-    /// <summary>
-    /// The kind of span (e.g. client, server).
-    /// </summary>
-    /// <seealso cref="Activity.Kind"/>
-    public const string SpanKind = "span.kind";
-
     /// <summary>
     /// The error message of an exception
     /// </summary>