+++ test/test-applications/integrations/TestApplication.Logs/Program.cs
@@ -14,7 +14,11 @@
 // limitations under the License.
 // </copyright>
 
+using System.Linq;
+using System.Net.Http;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Hosting.Server;
+using Microsoft.AspNetCore.Hosting.Server.Features;
 using Microsoft.Extensions.Hosting;
 
 namespace TestApplication.Logs;
@@ -23,7 +27,14 @@ public class Program
 {
     public static void Main(string[] args)
     {
-        CreateHostBuilder(args).Build().Run();
+        using var host = CreateHostBuilder(args).Build();
+        host.Start();
+
+        var server = (IServer)host.Services.GetService(typeof(IServer));
+        var addressFeature = server.Features.Get<IServerAddressesFeature>();
+        var address = addressFeature.Addresses.First();
+        using var httpClient = new HttpClient();
+        httpClient.GetAsync($"{address}/test").Wait();
     }
 
     public static IHostBuilder CreateHostBuilder(string[] args) =>
@@ -31,5 +42,6 @@ public class Program
             .ConfigureWebHostDefaults(webBuilder =>
             {
                 webBuilder.UseStartup<Startup>();
+                webBuilder.UseUrls($"http://127.0.0.1:0");
             });
 }